core-hash: xorror64: cast len to uint64_t before flipping bits
Fix difference between size_t and uint64_t on 32 bit systems causing
different end hash results to be produced.

+++ core-hash.c
@@ -562,7 +562,7 @@ uint32_t HOT OPTIMIZE3 stress_hash_mulxror64(const char *str, const size_t len)
  */
 uint32_t HOT OPTIMIZE3 stress_hash_xorror64(const char *str, const size_t len)
 {
-	register uint64_t hash = ~len;
+	register uint64_t hash = ~(uint64_t)len;
 	register size_t i;
 
 	for (i = len >> 3; i; ) {
